requirements.txt
marshmallow==3.19.0
marshmallow-enum==1.5.1
mypy-extensions==1.0.0
packaging==23.0
paho-mqtt==1.6.1
typing-inspect==0.8.0
typing_extensions==4.5.0
